backend/requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Migrate>=4.0.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
pymysql>=1.1.0
cryptography>=41.0.0
deepgram-sdk>=2.12.0
openai>=1.0.0
python-magic>=0.4.27
Werkzeug>=2.0.0
gunicorn>=21.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
isort>=5.0.0
mypy>=1.0.0 